requirements.txt
assemblyai==0.21.0
edge_tts==6.1.10
ffmpeg_python==0.2.0
Flask==3.0.2
Flask_Cors==4.0.0
g4f
google_api_python_client==2.117.0
httplib2==0.22.0
oauth2client==4.1.3
openai==0.28.0
playsound==1.3.0
protobuf==4.25.3
python-dotenv==1.0.1
Requests==2.31.0
sox==1.4.1
srt_equalizer==0.1.8
termcolor==2.4.0
nltk